+++ fastly/block_fastly_service_logging_elasticsearch.go
@@ -116,9 +116,10 @@ func (h *ElasticSearchServiceAttributeHandler) GetSchema() *schema.Schema {
 			ValidateDiagFunc: validateLoggingFormatVersion(),
 		}
 		blockAttributes["placement"] = &schema.Schema{
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Where in the generated VCL the logging call should be placed (ignored).",
+			Type:             schema.TypeString,
+			Optional:         true,
+			Description:      "Where in the generated VCL the logging call should be placed.",
+			ValidateDiagFunc: validateLoggingPlacement(),
 		}
 		blockAttributes["response_condition"] = &schema.Schema{
 			Type:        schema.TypeString,