+++ fastly/validators.go
@@ -33,6 +33,12 @@ func validateLoggingCompressionCodec() schema.SchemaValidateDiagFunc {
 	}, false))
 }
 
+func validateLoggingPlacement() schema.SchemaValidateDiagFunc {
+	return validation.ToDiagFunc(validation.StringInSlice([]string{
+		"none",
+	}, false))
+}
+
 func validateLoggingServerSideEncryption() schema.SchemaValidateDiagFunc {
 	return validation.ToDiagFunc(validation.StringInSlice([]string{
 		string(gofastly.S3ServerSideEncryptionAES),